+++ NEWS.md
@@ -1,6 +1,9 @@
 # `shinystan news`
 
 ### Version 2.2.0
+* Add optional argument `pars` to the `as.shinystan` method for stanfit objects,
+allowing a subset of parameters to be selected for inclusion in the resulting
+shinystan object.
 * Introduce `drop_parameters` function for removing parameters from a shinystan 
 object (useful for very large objects when you only want to look at a subset of
 parameters).